test(omo-senpi): make scheduler-lock deferral deterministic
